musiclstm/requirements.txt
lightning[pytorch-extra]==2.4.*
pandas==2.2.*
numpy<2.0
mido==1.3.*
matplotlib>=3.5.0
pyyaml>=6.0
tensorboard>=2.11.0
